PLAT-10723: Scrypto PoC
Add to move to Java 11 to use Scrypto as a composite build (put both
projects in the same folder for this to work).
Also upgraded Gradle because of the JMH plugin.

For now I implemented just message sending with a custom MessageService
that needs to access the pod's private APIs.
MessageML utils was also added to parse MML to PresentationML.